feat:cleanup(ex/test/factories): split factories into files
